refactor(swap): extract swap models, input collector, and extensions (#4865)
Move the self-contained state types and stateless helpers out of the
1000+ line SwapFormViewModel into dedicated swap/ package files:

- PreFlipState / ValidatedSwapInputs -> SwapFormModels.kt
- QuoteStateHolder -> QuoteStateHolder.kt
- formatFlippedAmount / updateSrc / MAX_DISPLAY_DECIMALS -> SwapFormExtensions.kt
- collectValidatedSwapInputs() -> SwapInputCollector.collect(), an injected
  collaborator that takes the current selections/fees and returns
  ValidatedSwapInputs or throws, leaving swap() a thin orchestrator and
  making the validation unit-testable without the ViewModel.
