+++ app/src/main/java/com/vultisig/wallet/ui/models/swap/SwapFormExtensions.kt
@@ -1,0 +1,38 @@
+package com.vultisig.wallet.ui.models.swap
+
+import com.vultisig.wallet.data.models.Address
+import com.vultisig.wallet.data.models.Chain
+import com.vultisig.wallet.ui.models.findCurrentSrc
+import com.vultisig.wallet.ui.models.firstSendSrc
+import com.vultisig.wallet.ui.models.send.SendSrc
+import java.math.BigDecimal
+import java.math.RoundingMode
+import kotlinx.coroutines.flow.MutableStateFlow
+
+private const val MAX_DISPLAY_DECIMALS = 8
+
+internal fun BigDecimal.formatFlippedAmount(tokenDecimals: Int? = null): String =
+    setScale(
+            (tokenDecimals ?: MAX_DISPLAY_DECIMALS).coerceAtMost(MAX_DISPLAY_DECIMALS),
+            RoundingMode.DOWN,
+        )
+        .stripTrailingZeros()
+        .toPlainString()
+
+internal fun MutableStateFlow<SendSrc?>.updateSrc(
+    selectedTokenId: String?,
+    addresses: List<Address>,
+    chain: Chain?,
+) {
+    val selectedSrcValue = value
+    value =
+        if (addresses.isEmpty()) {
+            null
+        } else {
+            if (selectedSrcValue == null) {
+                addresses.firstSendSrc(selectedTokenId, chain)
+            } else {
+                addresses.findCurrentSrc(selectedTokenId, selectedSrcValue)
+            }
+        }
+}
